chore: move display stuff in InfoDisplay
affected modules are: ClockDisplay, CpsDisplay, FpsDisplay and
TpsDisplay
